+++ GiftDecline/src/ModData.cs
@@ -5,7 +5,18 @@
 	/// <summary>Mod save game data.</summary>
 	internal class ModData
 	{
-		/// <summary>List of gift taste differences.</summary>
-		public Dictionary<string, Dictionary<string, int>> GiftTasteOverwrites { get; set; } = new Dictionary<string, Dictionary<string, int>>();
+		/// <summary>
+		/// List of gift taste differences.
+		/// Format: NpcName => itemId = taste.
+		/// </summary>
+		public Dictionary<string, Dictionary<string, int>> GiftTasteOverwrites { get; set; } =
+			new Dictionary<string, Dictionary<string, int>>();
+
+		/// <summary>
+		/// List of how many items a gift has been received.
+		/// Format: NpcName => itemId = amount.
+		/// </summary>
+		public Dictionary<string, Dictionary<string, int>> GiftTasteDeclineBuffer { get; set; } =
+			new Dictionary<string, Dictionary<string, int>>();
 	}
 }