+++ GiftDecline/src/SaveGameHelper.cs
@@ -1,6 +1,7 @@
 ﻿namespace GiftDecline
 {
 	using System.Collections.Generic;
+	using System.Linq;
 	using Common;
 	using StardewModdingAPI;
 	using StardewValley;
@@ -8,13 +9,13 @@
 	/// <summary>Methods to alter the save game data.</summary>
 	internal static class SaveGameHelper
 	{
+		/// <summary>Identifier to store and retrive this mod's data from the save game.</summary>
+		public const string Key = "GiftDecline";
+
 		private static IDataHelper dataHelper;
 
 		/// <summary>Save game data. This is only persisted for the main player.</summary>
 		public static ModData SaveState { get; set; }
-
-		/// <summary>Identifier to store and retrive this mod's data from the save game.</summary>
-		public static string Key { get; } = "GiftDecline";
 
 		/// <summary>Initialize the helper.</summary>
 		/// <param name="helper">Instance to use for reading from and writing to the save file.</param>
@@ -35,10 +36,20 @@
 			SaveState = dataHelper.ReadSaveData<ModData>(Key) ?? new ModData();
 		}
 
-		/// <summary>Clear the list of gift taste differences.</summary>
-		public static void ResetGiftTastes()
+		/// <summary>Send the save state to all peers.</summary>
+		public static void SyncWithPeers()
 		{
+			MultiplayerHelper.SendMessage(SaveState, Key);
+		}
+
+		/// <summary>Clear the list of gift taste differences.</summary>
+		public static void ResetSaveState()
+		{
+			Logger.Trace("Resetting save state");
+			NpcHelper.ResetGiftTastes();
+
 			SaveState.GiftTasteOverwrites = new Dictionary<string, Dictionary<string, int>>();
+			SaveState.GiftTasteDeclineBuffer = new Dictionary<string, Dictionary<string, int>>();
 		}
 
 		/// <summary>Set the gift taste overwrite value for an npc and item.</summary>
@@ -53,8 +64,33 @@
 			}
 
 			SaveState.GiftTasteOverwrites[npcName][itemId] = giftTaste;
+		}
 
-			MultiplayerHelper.SendMessage(SaveState, Key);
+		/// <summary>Update the buffer .</summary>
+		/// <param name="npc">NPC.</param>
+		/// <param name="item">Item.</param>
+		public static void HandleReceivedGift(NPC npc, Item item)
+		{
+			string npcName = npc.Name;
+			string itemId = item.ParentSheetIndex.ToString();
+
+			bool didBufferExceed = BumpGiftAmount(npcName, item);
+			if (didBufferExceed)
+			{
+				int newGiftTaste = NpcHelper.GetReduceGiftTaste(npc, item);
+				SetGiftTaste(npc.Name, itemId, newGiftTaste);
+			}
+
+			if (Context.IsMultiplayer)
+			{
+				// In single player, adjusting gift tastes is handled at the end of the day
+				// so the player can see the actual reaction to that gift in the social tab.
+				// But in Multiplayer, this delay causes some headaches when debugging some scenarios.
+				// So just apply it right away. This way all players are also in sync.
+				Apply();
+			}
+
+			SyncWithPeers();
 		}
 
 		/// <summary>Apply the current overwrites to the game.</summary>
@@ -64,10 +100,17 @@
 			{
 				NPC npc = Game1.getCharacterFromName(npcName);
 
-				foreach (string itemId in SaveState.GiftTasteOverwrites[npcName].Keys)
+				// .ToList because the save state is potentially "repaired" in this block
+				foreach (string itemId in SaveState.GiftTasteOverwrites[npcName].Keys.ToList())
 				{
 					Item item = new Object(int.Parse(itemId), 1);
-					NpcHelper.SetGiftTasteLevel(npc, item, SaveState.GiftTasteOverwrites[npcName][itemId]);
+					int currentTaste = SaveState.GiftTasteOverwrites[npcName][itemId];
+					int actualSetTaste = NpcHelper.SetGiftTasteLevel(npc, item, currentTaste);
+					if (actualSetTaste != currentTaste)
+					{
+						// fix save state, it is not valid with the current configuration
+						SaveState.GiftTasteOverwrites[npcName][itemId] = actualSetTaste;
+					}
 				}
 			}
 		}
@@ -90,9 +133,44 @@
 				return;
 			}
 
-			NpcHelper.ResetGiftTastes();
-			ResetGiftTastes();
+			ResetSaveState();
 			Logger.Info("Success");
+		}
+
+		/// <summary>Bump the amount this NPC has received this item.</summary>
+		/// <returns>Wether or not the gift exceeded the declining threshold.</returns>
+		private static bool BumpGiftAmount(string npcName, Item item)
+		{
+			string itemId = item.ParentSheetIndex.ToString();
+
+			if (!SaveState.GiftTasteDeclineBuffer.ContainsKey(npcName))
+			{
+				SaveState.GiftTasteDeclineBuffer.Add(npcName, new Dictionary<string, int>());
+			}
+
+			int existingAmount = 0;
+			if (SaveState.GiftTasteDeclineBuffer[npcName].ContainsKey(itemId))
+			{
+				existingAmount = SaveState.GiftTasteDeclineBuffer[npcName][itemId];
+			}
+
+			int newAmount = existingAmount + 1;
+			bool didExceedBuffer = newAmount >= ConfigHelper.Config.ReduceAfterXGifts;
+
+			if (didExceedBuffer)
+			{
+				string logExceeded = "Gifting threshold (" + ConfigHelper.Config.ReduceAfterXGifts + ") reached for ";
+				logExceeded += npcName + ", item " + NpcHelper.GetItemString(item);
+				Logger.Trace(logExceeded);
+				SaveState.GiftTasteDeclineBuffer[npcName].Remove(itemId);
+			}
+			else
+			{
+				Logger.Trace(npcName + " received item " + NpcHelper.GetItemString(item) + ". Total: " + newAmount);
+				SaveState.GiftTasteDeclineBuffer[npcName][itemId] = newAmount;
+			}
+
+			return didExceedBuffer;
 		}
 	}
 }