+++ GiftDecline/CHANGELOG.md
@@ -1,6 +1,9 @@
 ﻿## 1.1.0
 
 - Feature: [#2] Add option to cap gift taste reduction
+- Feature: [#3] Add option to delay gift taste reduction
+- Fix: Use configuration of the host in multiplayer
+- Fix: Clear gift taste changes when exiting to the title screen
 
 ## 1.0.1